requirements.txt
pdfplumber
pytesseract
Pillow
opencv-python
pandas
openpyxl
fastapi
uvicorn[standard]
